Variable type change

+++ notes_display.h
@@ -2,7 +2,7 @@
 #include "pitches.h"
 
 struct NoteDisplay {
-    unsigned long note;
+    uint16_t note;
     void (*digit1Func)();
     void (*digit2Func)();
     void (*digit3Func)();
@@ -116,7 +116,7 @@ const NoteDisplay noteMappings[] = {
     {g9, symbolBlank, letterG, number9},
 };
 
-void noteOnDisplay(unsigned long note) {
+void noteOnDisplay(uint16_t note) {
     for (const auto& entry : noteMappings) {
         if (entry.note == note) {
             digit1();
